+++ commands/server.go
@@ -19,6 +19,7 @@ import (
 	"os"
 	"strings"
 
+	"github.com/fsnotify/fsnotify"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"github.com/ystia/yorc/config"
@@ -38,21 +39,27 @@ func init() {
 }
 
 const (
-	environmentVariablePrefix = "YORC"
+	// EnvironmentVariablePrefix is the prefix used in Yorc commands parameters
+	// passed as environment variables
+	EnvironmentVariablePrefix = "YORC"
 )
 
 var (
-	tfConsulPluginVersion           = "tf Consul plugin version"
-	tfConsulPluginVersionConstraint = versionToConstraint("~>", tfConsulPluginVersion, "minor")
+	// TfConsulPluginVersion is the Terraform Consul plugin lowest supported version
+	TfConsulPluginVersion           = "tf Consul plugin version"
+	tfConsulPluginVersionConstraint = versionToConstraint("~>", TfConsulPluginVersion, "minor")
 
-	tfAWSPluginVersion           = "tf AWS plugin version"
-	tfAWSPluginVersionConstraint = versionToConstraint("~>", tfAWSPluginVersion, "minor")
+	// TfAWSPluginVersion is the Terraform AWS plugin lowest supported version
+	TfAWSPluginVersion           = "tf AWS plugin version"
+	tfAWSPluginVersionConstraint = versionToConstraint("~>", TfAWSPluginVersion, "minor")
 
-	tfOpenStackPluginVersion           = "tf OpenStack plugin version"
-	tfOpenStackPluginVersionConstraint = versionToConstraint("~>", tfOpenStackPluginVersion, "minor")
+	// TfOpenStackPluginVersion is the Terraform OpenStack plugin lowest supported version
+	TfOpenStackPluginVersion           = "tf OpenStack plugin version"
+	tfOpenStackPluginVersionConstraint = versionToConstraint("~>", TfOpenStackPluginVersion, "minor")
 
-	tfGooglePluginVersion           = "tf Google plugin version"
-	tfGooglePluginVersionConstraint = versionToConstraint("~>", tfGooglePluginVersion, "minor")
+	// TfGooglePluginVersion is the Terraform Google plugin lowest supported version
+	TfGooglePluginVersion           = "tf Google plugin version"
+	tfGooglePluginVersionConstraint = versionToConstraint("~>", TfGooglePluginVersion, "minor")
 )
 
 var ansibleConfiguration = map[string]interface{}{
@@ -115,14 +122,25 @@ var serverCmd = &cobra.Command{
 	},
 	RunE: func(cmd *cobra.Command, args []string) error {
 		log.Println("Using config file:", viper.ConfigFileUsed())
-		configuration := getConfig()
-		log.Debugf("Configuration :%+v", configuration)
 		shutdownCh := make(chan struct{})
-		return server.RunServer(configuration, shutdownCh)
+		return RunServer(shutdownCh)
 	},
 }
 
+// RunServer starts a Yorc Server
+func RunServer(shutdownCh chan struct{}) error {
+	configuration := GetConfig()
+	log.Debugf("Configuration :%+v", configuration)
+	return server.RunServer(configuration, shutdownCh)
+}
+
 func serverInitExtraFlags(args []string) {
+	InitExtraFlags(args, serverCmd)
+}
+
+// InitExtraFlags inits infrastructure and vault flags
+func InitExtraFlags(args []string, cmd *cobra.Command) {
+
 	resolvedServerExtraParams = []*serverExtraParams{
 		&serverExtraParams{
 			argPrefix:   "infrastructure_",
@@ -154,10 +172,10 @@ func serverInitExtraFlags(args []string) {
 					viperName = strings.Replace(strings.Replace(flagName, sep.argPrefix, sep.viperPrefix, 1), "_", ".", sep.subSplit)
 					if len(flagParts) == 1 {
 						// Boolean flag
-						serverCmd.PersistentFlags().Bool(flagName, false, "")
+						cmd.PersistentFlags().Bool(flagName, false, "")
 						viper.SetDefault(viperName, false)
 					} else {
-						serverCmd.PersistentFlags().String(flagName, "", "")
+						cmd.PersistentFlags().String(flagName, "", "")
 						viper.SetDefault(viperName, "")
 					}
 				} else {
@@ -165,16 +183,28 @@ func serverInitExtraFlags(args []string) {
 					flagName = strings.TrimLeft(args[i], "-")
 					viperName = strings.Replace(strings.Replace(flagName, sep.argPrefix, sep.viperPrefix, 1), "_", ".", sep.subSplit)
 					if len(args) > i+1 && !strings.HasPrefix(args[i+1], "--") {
-						serverCmd.PersistentFlags().String(flagName, "", "")
-						viper.SetDefault(viperName, "")
+
+						// Arguments ending wih a plural 's' are considered to
+						// be slices
+						if strings.HasSuffix(args[i], "s") && !strings.HasSuffix(args[i], "credentials") {
+							// May have already been defined as string slice
+							// flags can appear several times
+							if cmd.PersistentFlags().Lookup(flagName) == nil {
+								cmd.PersistentFlags().StringSlice(flagName, []string{}, "")
+								viper.SetDefault(viperName, []string{})
+							}
+						} else {
+							cmd.PersistentFlags().String(flagName, "", "")
+							viper.SetDefault(viperName, "")
+						}
 					} else {
 						// Boolean flag
-						serverCmd.PersistentFlags().Bool(flagName, false, "")
+						cmd.PersistentFlags().Bool(flagName, false, "")
 						viper.SetDefault(viperName, false)
 					}
 				}
 				// Add viper flag
-				viper.BindPFlag(viperName, serverCmd.PersistentFlags().Lookup(flagName))
+				viper.BindPFlag(viperName, cmd.PersistentFlags().Lookup(flagName))
 				sep.viperNames = append(sep.viperNames, viperName)
 			}
 		}
@@ -200,6 +230,13 @@ func initConfig() {
 	// If a config file is found, read it in.
 	if err := viper.ReadInConfig(); err != nil {
 		log.Println("Can't use config file:", err)
+	} else {
+		// Watch config to take into account config changes
+		viper.WatchConfig()
+		viper.OnConfigChange(func(e fsnotify.Event) {
+			log.Printf("Reloading config on config file %s change\n", e.Name)
+			viper.ReadInConfig()
+		})
 	}
 
 	// Deprecate Ansible and Consul flat keys if they are defined in
@@ -224,6 +261,7 @@ func setConfig() {
 	serverCmd.PersistentFlags().StringP("resources_prefix", "x", "", "Prefix created resources (like Computes and so on)")
 	serverCmd.PersistentFlags().Duration("wf_step_graceful_termination_timeout", config.DefaultWfStepGracefulTerminationTimeout, "Timeout to wait for a graceful termination of a workflow step during concurrent workflow step failure. After this delay the step is set on error.")
 	serverCmd.PersistentFlags().String("server_id", host, "The server ID used to identify the server node in a cluster.")
+	serverCmd.PersistentFlags().Bool("use_ssh_agent", true, "Whether or not allow using SSH-agent for SSH authentication on provisioned computes. Default is true. If false, compute credentials must provide a path to private key file.")
 
 	// Flags definition for Yorc HTTP REST API
 	serverCmd.PersistentFlags().Int("http_port", config.DefaultHTTPPort, "Port number for the Yorc HTTP REST API. If omitted or set to '0' then the default port number is used, any positive integer will be used as it, and finally any negative value will let use a random port.")
@@ -279,6 +317,7 @@ func setConfig() {
 	viper.BindPFlag("resources_prefix", serverCmd.PersistentFlags().Lookup("resources_prefix"))
 	viper.BindPFlag("wf_step_graceful_termination_timeout", serverCmd.PersistentFlags().Lookup("wf_step_graceful_termination_timeout"))
 	viper.BindPFlag("server_id", serverCmd.PersistentFlags().Lookup("server_id"))
+	viper.BindPFlag("use_ssh_agent", serverCmd.PersistentFlags().Lookup("use_ssh_agent"))
 
 	//Bind Flags Yorc HTTP REST API
 	viper.BindPFlag("http_port", serverCmd.PersistentFlags().Lookup("http_port"))
@@ -300,7 +339,7 @@ func setConfig() {
 	}
 
 	//Environment Variables
-	viper.SetEnvPrefix(environmentVariablePrefix)
+	viper.SetEnvPrefix(EnvironmentVariablePrefix)
 	viper.AutomaticEnv() // read in environment variables that match
 	viper.BindEnv("working_directory")
 	viper.BindEnv("plugins_directory")
@@ -315,6 +354,7 @@ func setConfig() {
 	viper.BindEnv("SSL_verify")
 	viper.BindEnv("resources_prefix")
 	viper.BindEnv("server_id")
+	viper.BindEnv("use_ssh_agent")
 
 	//Bind Consul environment variables flags
 	for key := range consulConfiguration {
@@ -343,6 +383,7 @@ func setConfig() {
 	viper.SetDefault("workers_number", config.DefaultWorkersNumber)
 	viper.SetDefault("wf_step_graceful_termination_timeout", config.DefaultWfStepGracefulTerminationTimeout)
 	viper.SetDefault("server_id", host)
+	viper.SetDefault("use_ssh_agent", true)
 
 	// Consul configuration default settings
 	for key, value := range consulConfiguration {
@@ -366,7 +407,8 @@ func setConfig() {
 
 }
 
-func getConfig() config.Configuration {
+// GetConfig gets configuration from viper
+func GetConfig() config.Configuration {
 	configuration := config.Configuration{}
 	err := viper.Unmarshal(&configuration)
 	if err != nil {
@@ -430,7 +472,25 @@ func addServerExtraInfraParams(cfg *config.Configuration, infraParam string) {
 		params = make(config.DynamicMap)
 		cfg.Infrastructures[paramParts[1]] = params
 	}
-	params.Set(paramParts[2], value)
+
+	// When the key/value pair is read from an environment variable, the value is
+	// read as a string. This needs to be changed if the variable is expected to
+	// be an array
+	if strings.HasSuffix(paramParts[2], "s") && !strings.HasSuffix(paramParts[2], "credentials") {
+		// value should be a slice
+		switch value.(type) {
+		case string:
+			vSlice := strings.Split(fmt.Sprint(value), ",")
+			for i, val := range vSlice {
+				vSlice[i] = strings.TrimSpace(val)
+			}
+			params.Set(paramParts[2], vSlice)
+		default:
+			params.Set(paramParts[2], value)
+		}
+	} else {
+		params.Set(paramParts[2], value)
+	}
 }
 
 func addServerExtraVaultParam(cfg *config.Configuration, vaultParam string) {
@@ -496,7 +556,7 @@ func toFlatKey(nestedKey string) string {
 // variable
 func toEnvVar(key string) string {
 
-	name := environmentVariablePrefix + "_" + toFlatKey(key)
+	name := EnvironmentVariablePrefix + "_" + toFlatKey(key)
 	return strings.ToUpper(name)
 }