+++ prov/terraform/google/generator.go
@@ -27,6 +27,7 @@ import (
 	"github.com/ystia/yorc/config"
 	"github.com/ystia/yorc/deployments"
 	"github.com/ystia/yorc/helper/consulutil"
+	"github.com/ystia/yorc/helper/sshutil"
 	"github.com/ystia/yorc/log"
 	"github.com/ystia/yorc/prov/terraform/commons"
 	"github.com/ystia/yorc/tosca"
@@ -37,11 +38,12 @@ const infrastructureName = "google"
 type googleGenerator struct {
 }
 
-func (g *googleGenerator) GenerateTerraformInfraForNode(ctx context.Context, cfg config.Configuration, deploymentID, nodeName, infrastructurePath string) (bool, map[string]string, []string, error) {
+func (g *googleGenerator) GenerateTerraformInfraForNode(ctx context.Context, cfg config.Configuration, deploymentID, nodeName, infrastructurePath string) (bool, map[string]string, []string, commons.PostApplyCallback, error) {
 	log.Debugf("Generating infrastructure for deployment with id %s", deploymentID)
+
 	cClient, err := cfg.GetConsulClient()
 	if err != nil {
-		return false, nil, nil, err
+		return false, nil, nil, nil, err
 	}
 	kv := cClient.KV()
 	nodeKey := path.Join(consulutil.DeploymentKVPrefix, deploymentID, "topology", "nodes", nodeName)
@@ -113,18 +115,20 @@ func (g *googleGenerator) GenerateTerraformInfraForNode(ctx context.Context, cfg
 	log.Debugf("inspecting node %s", nodeKey)
 	nodeType, err := deployments.GetNodeType(kv, deploymentID, nodeName)
 	if err != nil {
-		return false, nil, nil, err
+		return false, nil, nil, nil, err
 	}
 	outputs := make(map[string]string)
 	instances, err := deployments.GetNodeInstancesIds(kv, deploymentID, nodeName)
 	if err != nil {
-		return false, nil, nil, err
+		return false, nil, nil, nil, err
 	}
+
+	var sshAgent *sshutil.SSHAgent
 
 	for instNb, instanceName := range instances {
 		instanceState, err := deployments.GetInstanceState(kv, deploymentID, nodeName, instanceName)
 		if err != nil {
-			return false, nil, nil, err
+			return false, nil, nil, nil, err
 		}
 		if instanceState == tosca.NodeStateDeleting || instanceState == tosca.NodeStateDeleted {
 			// Do not generate something for this node instance (will be deleted if exists)
@@ -135,48 +139,66 @@ func (g *googleGenerator) GenerateTerraformInfraForNode(ctx context.Context, cfg
 		case "yorc.nodes.google.Compute":
 			instances, err = deployments.GetNodeInstancesIds(kv, deploymentID, nodeName)
 			if err != nil {
-				return false, nil, nil, err
+				return false, nil, nil, nil, err
 			}
 
-			err = g.generateComputeInstance(ctx, kv, cfg, deploymentID, nodeName, instanceName, instNb, &infrastructure, outputs)
+			err = g.generateComputeInstance(ctx, kv, cfg, deploymentID, nodeName, instanceName, instNb, &infrastructure, outputs, &cmdEnv, sshAgent)
 			if err != nil {
-				return false, nil, nil, err
+				return false, nil, nil, nil, err
 			}
 		case "yorc.nodes.google.Address":
 			err = g.generateComputeAddress(ctx, kv, cfg, deploymentID, nodeName, instanceName, instNb, &infrastructure, outputs)
 			if err != nil {
-				return false, nil, nil, err
+				return false, nil, nil, nil, err
 			}
 		case "yorc.nodes.google.PersistentDisk":
 			err = g.generatePersistentDisk(ctx, kv, cfg, deploymentID, nodeName, instanceName, instNb, &infrastructure, outputs)
 			if err != nil {
-				return false, nil, nil, err
+				return false, nil, nil, nil, err
 			}
 		case "yorc.nodes.google.PrivateNetwork":
 			err = g.generatePrivateNetwork(ctx, kv, cfg, deploymentID, nodeName, &infrastructure, outputs)
 			if err != nil {
-				return false, nil, nil, err
+				return false, nil, nil, nil, err
 			}
 		case "yorc.nodes.google.Subnetwork":
 			err = g.generateSubNetwork(ctx, kv, cfg, deploymentID, nodeName, &infrastructure, outputs)
 			if err != nil {
-				return false, nil, nil, err
+				return false, nil, nil, nil, err
 			}
 		default:
-			return false, nil, nil, errors.Errorf("Unsupported node type '%s' for node '%s' in deployment '%s'", nodeType, nodeName, deploymentID)
+			return false, nil, nil, nil, errors.Errorf("Unsupported node type '%s' for node '%s' in deployment '%s'", nodeType, nodeName, deploymentID)
 		}
 
 	}
 
+	// If ssh-agent has been created, it needs to be stopped after the infrastructure creation
+	// This is done with this callback
+	var postInstallCb func()
+	if sshAgent != nil {
+		postInstallCb = func() {
+			// Stop the sshAgent if used during provisioning
+			// Do not return any error if failure occured during this
+			err := sshAgent.RemoveAllKeys()
+			if err != nil {
+				log.Debugf("Warning: failed to remove all SSH agents keys due to error:%+v", err)
+			}
+			err = sshAgent.Stop()
+			if err != nil {
+				log.Debugf("Warning: failed to stop SSH agent due to error:%+v", err)
+			}
+		}
+	}
+
 	jsonInfra, err := json.MarshalIndent(infrastructure, "", "  ")
 	if err != nil {
-		return false, nil, nil, errors.Wrap(err, "Failed to generate JSON of terraform Infrastructure description")
+		return false, nil, nil, nil, errors.Wrap(err, "Failed to generate JSON of terraform Infrastructure description")
 	}
 
 	if err = ioutil.WriteFile(filepath.Join(infrastructurePath, "infra.tf.json"), jsonInfra, 0664); err != nil {
-		return false, nil, nil, errors.Wrapf(err, "Failed to write file %q", filepath.Join(infrastructurePath, "infra.tf.json"))
+		return false, nil, nil, nil, errors.Wrapf(err, "Failed to write file %q", filepath.Join(infrastructurePath, "infra.tf.json"))
 	}
 
 	log.Debugf("Infrastructure generated for deployment with id %s", deploymentID)
-	return true, outputs, cmdEnv, nil
+	return true, outputs, cmdEnv, postInstallCb, nil
 }