+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 ## UNRELEASED
 
+## 3.1.0-M7 (December 07, 2018)
+
+### DEPENDENCIES
+
+* The orchestrator requires now at least Ansible 2.7.2 (upgrade from 2.6.3 introduced in [GH-194](https://github.com/ystia/yorc/issues/194))
+
+### FEATURES
+
+* Allow to bootstrap a full stack Alien4Cloud/Yorc setup using yorc CLI ([GH-131](https://github.com/ystia/yorc/issues/131))
+
+### ENHANCEMENTS
+
+* Use ssh-agent to not write ssh private keys on disk ([GH-201](https://github.com/ystia/yorc/issues/201))
+
+### BUG FIXES
+
+* ConnectTo relationship not working for kubernetes topologies ([GH-212](https://github.com/ystia/yorc/issues/212))
+
 ## 3.1.0-M6 (November 16, 2018)
 
 ### FEATURES